+++ app/src/test/java/io/github/nodyssey/ui/account/PreferencesScreenTest.kt
@@ -1,7 +1,7 @@
 package io.github.nodyssey.ui.account
 
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.onRoot
 import io.github.nodyssey.ui.theme.NodysseyTheme