+++ app/src/test/java/io/github/nodyssey/ui/richtext/RichContentTest.kt
@@ -3,7 +3,7 @@ package io.github.nodyssey.ui.richtext
 import android.content.Context
 import androidx.compose.ui.test.assertHeightIsAtLeast
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.unit.dp